% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_predict.R
\name{CalPredictScore}
\alias{CalPredictScore}
\title{CalPredictScore}
\usage{
CalPredictScore(fit, new_data)
}
\arguments{
\item{fit}{A trained model object. Supported model types include "lognet", "glm", "svm.formula", "train", "glmboost", "plsRglmmodel", "rfsrc", "gbm", "xgb.Booster", and "naiveBayes".}

\item{new_data}{A data frame or matrix containing new data for prediction. Rows represent samples, and columns represent features.}
}
\value{
A named numeric vector of predicted values.
}
\description{
CalPredictScore
}
