% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{ExtractVar}
\alias{ExtractVar}
\title{Extract Variables Used in a Model}
\usage{
ExtractVar(fit)
}
\arguments{
\item{fit}{A fitted model object. The function supports various model types such as \code{lognet},
\code{glm}, \code{svm.formula}, \code{train}, \code{glmboost}, \code{plsRglmmodel}, \code{rfsrc}, \code{gbm}, \code{xgb.Booster},
and \code{naiveBayes}.}
}
\value{
A character vector of selected feature names used by the model. For models that do not
perform variable selection, the function returns all variables. \code{(Intercept)} or \code{Intercept}
terms are excluded from the output.
}
\description{
Extract Variables Used in a Model
}
\examples{
# Example with a random forest model (rfsrc)
library(randomForestSRC)
fit <- rfsrc(Species ~ ., data = iris)
variables <- ExtractVar(fit)
print(variables)

# Example with a generalized linear model (glm)
fit_glm <- glm(Species ~ Sepal.Length + Sepal.Width, data = iris, family = binomial)
variables_glm <- ExtractVar(fit_glm)
print(variables_glm)
}
