% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cal.R
\name{RunEval}
\alias{RunEval}
\title{RunEval}
\usage{
RunEval(
  fit,
  Test_set = NULL,
  Test_label = NULL,
  Train_set = NULL,
  Train_label = NULL,
  Train_name = NULL,
  cohortVar = "Cohort",
  classVar
)
}
\arguments{
\item{fit}{A trained model object.}

\item{Test_set}{A data frame or matrix containing the test data. Rows represent samples, and columns represent features.}

\item{Test_label}{A data frame containing the test metadata. This should include columns for cohort identifiers cohortVar and class labels classVar.}

\item{Train_set}{A data frame or matrix containing the training data. Rows represent samples, and columns represent features.}

\item{Train_label}{A data frame containing the training metadata. This should include the class labels classVar.}

\item{Train_name}{A string specifying the cohort name for the training data. If not provided, defaults to "Training".}

\item{cohortVar}{A string specifying the column name in Test_label and Train_label that identifies cohorts}

\item{classVar}{A string specifying the column name in Test_label and Train_label that identifies the target class labels.}
}
\description{
RunEval
}
