% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{RunML}
\alias{RunML}
\title{RunML}
\usage{
RunML(method, Train_set, Train_label, mode = "Model", classVar)
}
\arguments{
\item{method}{A string specifying the method used to create the model. Must be a supported method.}

\item{Train_set}{A matrix representing the training data. Rows correspond to samples, and columns correspond to genes.}

\item{Train_label}{Rows correspond to samples, and columns represent phenotypes. The phenotype should be encoded as 0 or 1.}

\item{mode}{A string with two possible values: 'Model' (default) or 'Variable'. 'Model' trains a model, while 'Variable' filters features.}

\item{classVar}{A string specifying the column name in Train_label to use as the target phenotype.}
}
\value{
A trained model or filtered features, depending on the selected mode.
}
\description{
RunML
}
