% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_plot.R
\name{SimpleHeatmap}
\alias{SimpleHeatmap}
\title{SimpleHeatmap}
\usage{
SimpleHeatmap(
  Cindex_mat,
  avg_Cindex,
  CohortCol,
  barCol,
  cellwidth = 1,
  cellheight = 0.5,
  cluster_columns,
  cluster_rows
)
}
\arguments{
\item{Cindex_mat}{A matrix containing C-index values for visualization. Rows correspond to items, and columns correspond to cohorts.}

\item{avg_Cindex}{A numeric vector representing the average C-index values for rows in Cindex_mat.}

\item{CohortCol}{A named vector or list specifying the colors for each cohort in Cindex_mat. Names must match the column names of Cindex_mat.}

\item{barCol}{A vector specifying the fill colors for the bar plot representing avg_Cindex. Each element corresponds to a row in avg_Cindex.}

\item{cellwidth}{A numeric value specifying the width of each cell in the heatmap (default: 1).}

\item{cellheight}{A numeric value specifying the height of each cell in the heatmap (default: 0.5).}

\item{cluster_columns}{A logical value or clustering method specifying whether to cluster the columns of the heatmap. Defaults to user-specified input.}

\item{cluster_rows}{A logical value or clustering method specifying whether to cluster the rows of the heatmap. Defaults to user-specified input.}
}
\value{
A heatmap object visualizing the C-index matrix, annotated with cohort labels and bar plots for row-level averages.
}
\description{
SimpleHeatmap
}
